Patched the code generator to support Windows x64 calling convention. Patched incorrect pointer to long casts (long on Windows x64 is 32bits!).

+++ qemu/cpu-all.h
@@ -651,8 +651,8 @@ extern int have_guest_base;
 #else /* !CONFIG_USER_ONLY */
 /* NOTE: we use double casts if pointers and target_ulong have
    different sizes */
-#define saddr(x) (uint8_t *)(long)(x)
-#define laddr(x) (uint8_t *)(long)(x)
+#define saddr(x) (uint8_t *)(uintptr_t)(x)
+#define laddr(x) (uint8_t *)(uintptr_t)(x)
 #endif
 
 #define ldub_raw(p) ldub_p(laddr((p)))
